Lint fixes

+++ core/lib/config/src/configs/base_token_adjuster.rs
@@ -7,7 +7,7 @@ pub const DEFAULT_INTERVAL_MS: u64 = 30_000;
 
 #[derive(Debug, Clone, PartialEq, Deserialize)]
 pub struct BaseTokenAdjusterConfig {
-    /// How often to poll external APIs for a new ETH<->BaseToken price.
+    /// How often to poll external APIs for a new ETH<->Base-Token price.
     pub price_polling_interval_ms: Option<u64>,
 }
 
@@ -20,8 +20,8 @@ impl BaseTokenAdjusterConfig {
 
     pub fn price_polling_interval(&self) -> Duration {
         match self.price_polling_interval_ms {
-            Some(interval) => Duration::from_millis(interval as u64),
-            None => Duration::from_millis(DEFAULT_INTERVAL_MS as u64),
+            Some(interval) => Duration::from_millis(interval),
+            None => Duration::from_millis(DEFAULT_INTERVAL_MS),
         }
     }
 }
